Cleaning up the code
- Reorganized the code a bit so it's easier to understand what is where.
- Also cleaned up the stats command as it was the first one built and
had several iterations of decisions.
